Updated and separated the commands.js files
