+++ cypress/support/commands.js
@@ -123,127 +123,6 @@ Cypress.Commands.add('searchableDropdown', (labelFor, labelText, item, options =
   cy.wait(300);
 });
 
-Cypress.Commands.add('selectCallCentreAndProvinces', (userFixture) => {
-  const { role, call_centre, province } = userFixture;
-
-  // Select Call Centre using stable logic
-  cy.searchableDropdown('call_center_id', 'Call Centre', call_centre, { clear: true });
-
-  // Wait for Livewire to finish updating the provinces
-  if (call_centre === 'Head Office') {
-    cy.wait(1000);
-  }
-
-  // If Head Office user, select ALL provinces using existing dropdown command
-  if (role === 'National Admin (Head Office)' && call_centre === 'Head Office') {
-    const provinces = [
-      'Eastern Cape', 'Free State', 'Gauteng', 'KwaZulu-Natal',
-      'Limpopo', 'Mpumalanga', 'Northern Cape', 'North West', 'Western Cape'
-    ];
-
-    provinces.forEach((prov, index) => {
-      cy.searchableDropdown('province_ids', 'Province', prov);
-    });
-
-  } else {
-    // Normal case: just select one province
-    cy.searchableDropdown('province_ids', 'Province', province, { clear: true });
-  }
-});
-
-// Creating a user on the system
-Cypress.Commands.add('generateUserFixture', () => {
-  const allProvinces = [
-    'Eastern Cape',
-    'Free State',
-    'Gauteng',
-    'International',
-    'KwaZulu-Natal',
-    'Limpopo',
-    'Mpumalanga',
-    'North West',
-    'Northern Cape',
-    'Unknown',
-    'Western Cape'
-  ];
-
-  const roleToRegionMap = {
-    'Eastern Cape Call Centre': 'Eastern Cape',
-    'Free State Call Centre': 'Free State',
-    'Gauteng Call Centre': 'Gauteng',
-    'Kwazulu-Natal Call Centre': 'KwaZulu-Natal',
-    'Limpopo Call Centre': 'Limpopo',
-    'Mpumalanga Call Centre': 'Mpumalanga',
-    'Northern Cape Call Centre': 'Northern Cape',
-    'North West Call Centre': 'North West',
-    'Western Cape Call Centre': 'Western Cape',
-    'International Call Centre': 'International'
-  };
-
-  const roles = [
-    'User (Viewer)',
-    'Provincial Admin',
-    'Data Capturer',
-    'National Admin (Head Office)'
-  ];
-
-  const callCentres = Object.keys(roleToRegionMap);
-  callCentres.push('Head Office');
-
-  let userFixture = {};
-
-  return cy.generateSAID().then((saId) => {
-    userFixture.id_number = saId;
-
-    return cy.generateSACellphone().then((cellphoneNumber) => {
-      userFixture.cell = cellphoneNumber;
-
-      return cy.generateLanguage().then((language) => {
-        userFixture.language = language;
-
-        return cy.task('generateUser').then((user) => {
-          userFixture.name = user.name;
-          userFixture.surname = user.surname;
-          userFixture.email = user.email;
-
-          // Assign a role
-          const selectedRole = roles[Math.floor(Math.random() * roles.length)];
-          userFixture.role = selectedRole;
-
-          if (selectedRole === 'National Admin (Head Office)') {
-            userFixture.call_centre = 'Head Office';
-            userFixture.provinces = allProvinces;
-          } else {
-            const entries = Object.entries(roleToRegionMap);
-            const [callCentre, province] = entries[Math.floor(Math.random() * entries.length)];
-            userFixture.call_centre = callCentre;
-            userFixture.province = province;
-            userFixture.provinces = [province];
-          }
-
-          // Assertions for sanity check
-          expect(userFixture.id_number).to.match(/^\d{13}$/);
-          expect(userFixture.cell).to.match(/^0\d{9}$/);
-          expect(['Afrikaans', 'English']).to.include(userFixture.language);
-          expect(userFixture.name).to.be.a('string');
-          expect(userFixture.surname).to.be.a('string');
-          expect(userFixture.email).to.match(/^[^\s@]+@[^\s@]+\.[^\s@]+$/);
-
-          // Debug logs
-          cy.log(JSON.stringify(userFixture, null, 2));
-
-          // Save to fixture file
-          return cy.writeFile('cypress/fixtures/user.json', userFixture).then(() => {
-            cy.log('User fixture has been written to user.json');
-            return cy.wrap(userFixture);
-          });
-        });
-      });
-    });
-  });
-});
-
-
 // Generating SA ID
 Cypress.Commands.add('generateSAID', () => {
   const generateSAID = () => {
@@ -331,125 +210,6 @@ Cypress.Commands.add('generateLanguage', () => {
   };
 
   return generateLanguage();
-});
-
-const permissionMap = { 
-  "Members": ['view', 'create', 'edit', 'delete', 'search', 'freeze', 'rejoin'],
-  "Members-details": ['view', 'edit'],
-  "Members-addresses": ['view', 'create', 'edit', 'delete'],
-  "Members-products": ['view', 'create', 'edit', 'delete'],
-  "Members-notes": ['view', 'create'],
-  "Members-documents": ['view', 'create', 'delete'],
-  "Members-communication": ['view'],
-  "Members-audit": ['view'],
-  "Waiting-rooms": ['view', 'edit'],
-  "Tickets": ['view', 'create', 'edit', 'delete'],
-  "Procomm": ['view', 'settings', 'approvers'],
-  "Reports": ['view', 'create', 'edit', 'delete'],
-  "Voters-roll": ['view', 'create', 'edit', 'delete', 'reports'],
-  "Voters-options": ['view', 'create', 'edit', 'delete'],
-  "Users": ['view', 'create', 'edit', 'delete', 'impersonate'],
-  "User-security": ['view', 'edit'],
-  "User-authentication-log": ['view'],
-  "User-activities": ['view'],
-  "User-call-center": ['edit'],
-  "User-province": ['edit'],
-  "Roles": ['view', 'create', 'edit', 'delete'],
-  "Permissions": ['view', 'edit'],
-  "Administration": ['view'],
-  "Branches": ['view', 'create', 'edit', 'delete'],
-  "Calendar": ['view', 'create'],
-  "Notices": ['view', 'create', 'edit', 'delete'],
-  "Member-roles": ['view', 'create', 'edit', 'delete'],
-  "Member-role-levels": ['view', 'create', 'edit', 'delete'],
-  "Member-role-categories": ['view', 'create', 'edit', 'delete'],
-  "Member-role-groups": ['view', 'create', 'edit', 'delete'],
-  "Member-join-date": ['edit'],
-  "Genders": ['view', 'create', 'edit', 'delete'],
-  "Occupations": ['view', 'create', 'edit', 'delete'],
-  "Products": ['view', 'create', 'edit', 'delete'],
-  "Product-types": ['view', 'create', 'edit', 'delete'],
-  "Payment-methods": ['view', 'create', 'edit', 'delete'],
-  "Tags": ['view', 'create', 'edit', 'delete'],
-  "Note-types": ['view', 'create', 'edit', 'delete'],
-  "Document-types": ['view', 'create', 'edit'],
-  "Member-type": ['view', 'create', 'edit', 'delete'],
-  "Product-settings": ['view', 'edit'],
-  "Banking-details": ['view', 'edit', 'create', 'delete']
-};
-
-// Generating Role Fixture
-Cypress.Commands.add('generateRoleFixture', () => {
-  const roleFixture = {};
-
-  const randomSuffix = Math.random().toString(36).substring(2, 7).toUpperCase();
-  roleFixture.name = `CyRole-${randomSuffix}`;
-
-  roleFixture.permissions = [];
-
-  // Loop through each section and randomly pick some permissions
-  Object.entries(permissionMap).forEach(([section, perms]) => {
-    const selectedPerms = perms.filter(() => Math.random() < 0.3);
-    selectedPerms.forEach(permission => {
-      roleFixture.permissions.push({ section, permission });
-    });
-  });
-
-  cy.log(`Generated role fixture: ${roleFixture.name}`);
-
-  return cy.writeFile('cypress/fixtures/role.json', roleFixture).then(() => {
-    return cy.wrap(roleFixture);
-  });
-});
-
-// Accordion expand command
-Cypress.Commands.add('expandAccordionInContainer', (label) => {
-  cy.get('div.max-w-3xl.mx-auto.divide-y.divide-gray-200')
-    .contains('span', label)
-    .closest('button')
-    .click({ force: true });
-});
-
-// Toggle permission in accordion
-Cypress.Commands.add('togglePermissionInAccordion', (accordionLabel, ...permissions) => {
-  cy.log(`Expanding accordion: ${accordionLabel}`);
-
-  cy.get('div.max-w-3xl.mx-auto.divide-y.divide-gray-200')
-    .contains('span', accordionLabel)
-    .closest('button')
-    .as('accordionButton')
-    .scrollIntoView() // Ensure button is visible
-    .click({ force: true });
-
-  cy.wait(500);
-
-  // Scope to the expanded accordion container
-  cy.get('@accordionButton')
-    .parent() 
-    .next()   
-    .within(() => {
-      permissions.forEach(permissionLabel => {
-        cy.get('label').each(($label) => {
-          const labelText = $label.text().trim().toLowerCase();
-          if (labelText === permissionLabel.toLowerCase()) {
-            cy.wrap($label)
-              .scrollIntoView() 
-              .invoke('attr', 'for')
-              .then((id) => {
-                cy.get(`#${id}`)
-                  .scrollIntoView() 
-                  .then(($checkbox) => {
-                    if ($checkbox.prop('checked')) {
-                      cy.wrap($checkbox).uncheck({ force: true });
-                    } else {
-                      cy.wrap($checkbox).check({ force: true });
-                    }
-                  });
-              });
-          }
-        });
-      });
-    });
 });
 
 Cypress.Commands.add('collectTableHeaders', () => {
@@ -601,41 +361,4 @@ Cypress.Commands.add('clickAllPaginationPages', () => {
   }
 
   clickNextIfExists();
-});
-
-Cypress.Commands.add('branchDropdown', (labelFor, labelText, item, options = { clear: false }) => {
-  const container = cy.get(`div[form-wrapper="${labelFor}"]`);
-
-  if (options.clear) {
-    container
-      .find('div[name="form.wrapper.container.append"]')
-      .find('button')
-      .first()
-      .click({ force: true });
-    cy.wait(500);
-  }
-
-  // Open dropdown
-  cy.contains('label', labelText)
-    .should('be.visible')
-    .click();
-
-  cy.get('.max-h-80:visible')
-    .first()
-    .should('be.visible')
-    .within(() => {
-      if (item === '__select_first__') {
-        // Click the very first visible option
-        cy.get('li[tabindex="-1"], li[tabindex="0"]')
-          .filter(':visible')
-          .first()
-          .find('div')
-          .first()
-          .click({ force: true });
-      } else {
-        cy.contains('div', item)
-          .should('be.visible')
-          .click({ force: true });
-      }
-    });
 });
